chore(tests): splitted schema into modules
