+++ src/schema/modules/mocked-type.ts
@@ -1,0 +1,7 @@
+export const typeDef = `
+    # Type which should have values mocked
+    type MockedType {
+        mockedFirstName: String
+        mockedInt: Int
+    }
+`;
