+++ src/schema/modules/some-type.ts
@@ -1,0 +1,15 @@
+export const typeDef = `
+type SomeType {
+    testInt: Int
+    testFloat: Float
+    fixedString: String
+}
+`;
+
+export const resolver = {
+    SomeType: {
+        fixedString() {
+            return "fixed.";
+        },
+    },
+};
